backend/requirements.txt
asgiref==3.3.4
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.12.0
django-mathfilters==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
gunicorn==20.0.4
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.1.1
pycodestyle==2.8.0
PyJWT==2.4.0
python-dateutil==2.8.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.1
sweetify==2.0.0
toml==0.10.2
typing-extensions==3.10.0.0
waitress==2.1.2
whitenoise==5.2.0
XlsxWriter==3.0.2
